switch to app pane and only open on btn click
